requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
backoff==2.2.1
beautifulsoup4==4.12.2
blinker==1.6.3
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
cohere==4.31
Deprecated==1.2.14
fastavro==1.8.2
ffmpeg==1.4
Flask==3.0.0
frozenlist==1.4.0
future==0.18.3
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
lyricsgenius==3.0.1
MarkupSafe==2.1.3
multidict==6.0.4
networkx==2.8.8
numpy==1.24.4
python-dotenv==1.0.0
python-tsp==0.4.0
pytube==15.0.0
redis==5.0.1
requests==2.31.0
scikit-learn==1.3.1
scipy==1.10.1
six==1.16.0
soupsieve==2.5
spotipy==2.23.0
tabulate==0.8.10
threadpoolctl==3.2.0
tsplib95==0.7.1
urllib3==2.0.7
Werkzeug==3.0.0
wrapt==1.15.0
yarl==1.9.2
zipp==3.17.0
gunicorn==21.2.0
packaging==23.2
